地图系统三层渲染升级

+++ Tools/资源编辑器/MillenniumResEditor/Model/MapTile.cs
@@ -5,10 +5,10 @@ namespace MillenniumResEditor.Model
     /// </summary>
     public class MapTile
     {
-        /// <summary>底层瓦片索引</summary>
-        public byte Low { get; set; }
-        /// <summary>高层瓦片索引</summary>
-        public byte High { get; set; }
+        /// <summary>底层瓦片索引（16位，支持0-65535）</summary>
+        public ushort Low { get; set; }
+        /// <summary>高层瓦片索引（16位，支持0-65535）</summary>
+        public ushort High { get; set; }
         /// <summary>属性 0=通行 1=阻挡 2=事件区</summary>
         public byte Attr { get; set; }
     }
